Pull out branch list from branch card
Passing it as a snippet gives the branch card fewer dependencies, and makes it easier to customise.
